+++ src/main/java/com/lucendar/strm/common/utils/StrmUtils.java
@@ -1,5 +1,9 @@
 package com.lucendar.strm.common.utils;
 
+import java.nio.ByteBuffer;
+import java.util.Base64;
+import java.util.UUID;
+
 public class StrmUtils {
 
     private static final char[] HEX_CHARS = "0123456789abcdef".toCharArray();
@@ -17,4 +21,25 @@ public class StrmUtils {
 
         return new String(r);
     }
+
+    public static String timeCodedId(long time) {
+        UUID uuid = UUID.randomUUID();
+        byte[] bytes = new byte[16 + 8];
+        ByteBuffer bb = ByteBuffer.wrap(bytes);
+
+        bb.putLong(time);
+        bb.putLong(uuid.getMostSignificantBits());
+        bb.putLong(uuid.getLeastSignificantBits());
+
+        return Base64.getUrlEncoder().encodeToString(bytes);
+    }
+
+    public static long extractTimeFromId(String id) {
+        byte[] bytes = Base64.getUrlDecoder().decode(id);
+        ByteBuffer bb = ByteBuffer.wrap(bytes);
+        return bb.getLong();
+    }
+
+
+
 }